+++ packages/react-magma-docs/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### 2.1.3-alpha.5 (2021-01-25)
+
+**Note:** Version bump only for package react-magma-docs
+
+
+
+
+
+### 2.1.3-alpha.4 (2021-01-25)
+
+**Note:** Version bump only for package react-magma-docs
+
+
+
+
+
+### 2.1.3-alpha.3 (2021-01-22)
+
+**Note:** Version bump only for package react-magma-docs
+
+
+
+
+
+### 2.1.3-alpha.2 (2021-01-22)
+
+**Note:** Version bump only for package react-magma-docs
+
+
+
+
+
+### 2.1.3-alpha.1 (2021-01-21)
+
+**Note:** Version bump only for package react-magma-docs
+
+
+
+
+
+### 2.1.3-alpha.0 (2021-01-20)
+
+**Note:** Version bump only for package react-magma-docs
+
+
+
+
+
 ### 2.1.2 (2021-01-20)
 
 **Note:** Version bump only for package react-magma-docs